# api/openapi.yaml
openapi: 3.0.3
info:
  title: PayFlow API
  version: 1.0.0
  description: Minimal OpenAPI contract for the PayFlow payment gateway API.
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessage"
  /ready:
    get:
      summary: Readiness check
      responses:
        "200":
          description: Service is ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessage"
  /metrics:
    get:
      summary: Prometheus metrics
      responses:
        "200":
          description: Prometheus exposition format
          content:
            text/plain:
              schema:
                type: string
  /api/v1/transactions:
    post:
      summary: Create transaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessTransactionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/transactions/{id}:
    get:
      summary: Get transaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TransactionID"
      responses:
        "200":
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessTransactionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/transactions/{id}/refund:
    post:
      summary: Refund transaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TransactionID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundRequest"
      responses:
        "200":
          description: Transaction refunded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessTransactionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/transactions/{id}/events:
    get:
      summary: List transaction events
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TransactionID"
      responses:
        "200":
          description: Transaction events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEventsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/merchants/{id}/balance:
    get:
      summary: Get merchant balance
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/MerchantID"
      responses:
        "200":
          description: Merchant balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMerchantBalanceResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/merchants/{id}/transactions:
    get:
      summary: List merchant transactions
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/MerchantID"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Paginated transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessTransactionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    TransactionID:
      name: id
      in: path
      required: true
      schema:
        type: string
    MerchantID:
      name: id
      in: path
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: false
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    GenericMessage:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CreateTransactionRequest:
      type: object
      properties:
        amount_cents:
          type: integer
          format: int64
          minimum: 1
        currency:
          type: string
          example: CAD
        description:
          type: string
      required:
        - amount_cents
        - currency
    RefundRequest:
      type: object
      properties:
        reason:
          type: string
    Transaction:
      type: object
      properties:
        id:
          type: string
        merchant_id:
          type: string
        amount_cents:
          type: integer
          format: int64
        currency:
          type: string
        status:
          type: string
        idempotency_key:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TransactionEvent:
      type: object
      properties:
        id:
          type: integer
        transaction_id:
          type: string
        event_type:
          type: string
        from_status:
          type: string
        to_status:
          type: string
        created_at:
          type: string
          format: date-time
    MerchantBalance:
      type: object
      properties:
        merchant_id:
          type: string
        balance_cents:
          type: integer
          format: int64
        currency:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    SuccessTransactionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/Transaction"
    SuccessEventsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/TransactionEvent"
    SuccessMerchantBalanceResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/MerchantBalance"
    SuccessTransactionListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        meta:
          type: object
          properties:
            total:
              type: integer
            limit:
              type: integer
            offset:
              type: integer